ステージ２のマップです。

+++ ES Game Library/GameMain.h
@@ -69,7 +69,7 @@ private:
 	float jumpspeed, jumptime, zahyou;
 	float kunai_x, kunai_y, kunai2_x, kunai2_y;//攻撃用クナイ
 	float  floor2_1x,floor2_1y;
-	int frame, time;
+	float frame, time;
 
 	// 関数宣言
